feat(arena/tui): InspectPage + extract inspect collector/renderer package
