+++ tools/arena/inspect/test_init.go
@@ -1,0 +1,13 @@
+package inspect
+
+import (
+	"testing"
+
+	"github.com/AltairaLabs/PromptKit/pkg/config"
+)
+
+func init() {
+	if testing.Testing() {
+		config.SchemaValidationDisabled.Store(true)
+	}
+}
